Add files via upload
related_parties.py: python/postgresql code identifying the counterparties for a Trump transaction programmatically. Script queries a Trump transaction documentID from the listing of all transactions. This results in a listing of all the parties for that documentID. The script then continues on to the next documentID until all have been queried. The results are stored in a new table.